// src/@saas-boilerplate/providers/payment/types.ts
export type CyclePeriod = 'day' | 'week' | 'month' | 'year'

/**
 * Represents the status of a subscription.
 * @enum {string}
 */
export type SubscriptionStatus =
  | 'active'
  | 'canceled'
  | 'incomplete'
  | 'incomplete_expired'
  | 'past_due'
  | 'unpaid'
  | 'trialing'
// ...
export type Customer = {
  id: string
  providerId: string
  organizationId: string
  name: string
  email: string
  metadata?: Record<string, unknown>

  subscription?: {
    id: string
    providerId: string
    status: SubscriptionStatus
    trialDays: number | null

    plan: {
      id: string
      providerId: string
      slug: string
      name: string
      description: string
      metadata: PlanMetadata
      price: {
        id: string
        providerId: string
        slug: string
        amount: number
        currency: string
        interval: CyclePeriod
      }
    }

    usage: Usage[]

    createdAt?: Date
    updatedAt?: Date
  }

  createdAt?: Date
  updatedAt?: Date
}
// ...
export type PlanFeature = {
  slug: string
  name: string
  enabled: boolean
  description?: string
  table?: string
  limit?: number
  cycle?: CyclePeriod
}

/**
 * Represents metadata associated with a subscription plan.
 * @interface PlanMetadata
 * @property {PlanFeature[]} features - List of features included in this plan
 */
export type PlanMetadata = {
  features: PlanFeature[]
}
// ...
export type Usage = {
  slug: string
  name: string
  description?: string
  usage: number
  limit: number
  lastReset?: Date
  nextReset?: Date
  cycle: CyclePeriod
}